[ADD] Amount To Long

+++ data/src/main/java/com/woowahan/accountbook/data/datasource/HistoryDataSource.kt
@@ -10,7 +10,7 @@ interface HistoryDataSource {
     suspend fun getAllHistoriesByMonthAndType(firstDayOfMonth: Long, firstDayOfNextMonth: Long, type: String): List<History>
     suspend fun createHistoryTable()
     suspend fun dropHistoryTable()
-    suspend fun insertHistory(date: Long, amount: Int, content: String, category: Category, paymentMethod: PaymentMethod?)
-    suspend fun updateHistory(id: Int, date: Long, amount: Int, content: String, category: Category, paymentMethod: PaymentMethod?)
+    suspend fun insertHistory(date: Long, amount: Long, content: String, category: Category, paymentMethod: PaymentMethod?)
+    suspend fun updateHistory(id: Int, date: Long, amount: Long, content: String, category: Category, paymentMethod: PaymentMethod?)
     suspend fun deleteHistory(id: Int)
 }
