+++ domain/src/main/java/com/woowahan/accountbook/domain/usecase/history/UpdateHistoryUseCase.kt
@@ -12,7 +12,7 @@ import java.lang.Exception
 class UpdateHistoryUseCase(
     private val repository: HistoryRepository
 ) {
-    operator fun invoke(id: Int, date: Long, amount: Int, content: String, category: Category, paymentMethod: PaymentMethod?) = flow {
+    operator fun invoke(id: Int, date: Long, amount: Long, content: String, category: Category, paymentMethod: PaymentMethod?) = flow {
         try {
             repository.updateHistory(id, date, amount, content, category, paymentMethod)
             emit(Result.Success(Unit))
